Harvester: Refactor Novnc
