+++ shell/assets/translations/en-us.yaml
@@ -5865,7 +5865,7 @@ harvester:
         down: No events in the past hour
       console:
         down: This Virtual Machine is down. Please start it to access its console.
-        shortKeys: Short Keys
+        shortKeys: Shortcut Keys
 
   volume:
     label: Volumes
