+++ shell/components/novnc/NovncConsole.vue
