+++ shell/pages/c/_cluster/harvester/console/_uid/vnc.vue
@@ -1,6 +1,6 @@
 <script>
 import { HCI } from '@shell/config/types';
-import NovncConsoleWrapper from '@shell/components/NovncConsoleWrapper';
+import NovncConsoleWrapper from '@shell/components/novnc/NovncConsoleWrapper';
 import Loading from '@shell/components/Loading';
 
 export default {
@@ -46,7 +46,7 @@ export default {
 </template>
 
 <style>
-HTML, BODY, MAIN, #__nuxt, #__layout, #app {
+HTML, BODY, MAIN, #__nuxt, #__layout, #app, .vm-console, .vm-console > DIV, .vm-console > DIV > DIV {
   height: 100%;
 }
 </style>
